ネットワークの処理を切り出す
